+++ packages/auto_mappr/test/integration/when_source_is_null_test.dart
@@ -1,0 +1,33 @@
+import 'package:test/test.dart';
+
+import 'fixture/when_source_is_null.dart' as fixture;
+
+void main() {
+  late final fixture.Mappr mappr;
+
+  setUpAll(() {
+    mappr = fixture.Mappr();
+  });
+
+  test('whenSourceIsNull with static value ', () {
+    final source = fixture.SourceValue(null);
+    final target = mappr.convert<fixture.SourceValue, fixture.Target>(source);
+
+    expect(target.name, 'static');
+  });
+
+  test('whenSourceIsNull with static function in Mappr', () {
+    final source = fixture.SourceStatic(null);
+    final target = mappr.convert<fixture.SourceStatic, fixture.Target>(source);
+
+    expect(target.name, 'whenSourceIsNull');
+  });
+
+  test('whenSourceIsNull with top level function', () {
+    final source = fixture.SourceTopLevel(null);
+    final target =
+        mappr.convert<fixture.SourceTopLevel, fixture.Target>(source);
+
+    expect(target.name, 'whenSourceIsNullTopLevel');
+  });
+}