+++ packages/auto_mappr/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+- Add when `whenSourceIsNull` support to enum mapping. (By @DevNico) [#59](https://github.com/netglade/auto_mappr/pull/59)
+- Add enum mapping fallback support for missing fields in the Target using `whenSourceIsNull`. (By @DevNico) [#59](https://github.com/netglade/auto_mappr/pull/59)
+- Fix top-level functions now work with `whenSourceIsNull`. (By @DevNico) [#59](https://github.com/netglade/auto_mappr/pull/59)
+
 ## 1.3.1
 - Add documentation for enums and converting iterables. [#52](https://github.com/netglade/auto_mappr/pull/52)
 - Fix generator when multiple annotations are used. [#51](https://github.com/netglade/auto_mappr/pull/51)